Warning on 64-bit architecture fixed
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@15417 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/loggerta/logger_ta.c
@@ -430,7 +430,11 @@ log_get_message(uint32_t length, uint8_t *buffer)
                 LGR_32_TO_NET(tmp, tmp_buf);
                 tmp_buf += sizeof(uint32_t);
 
-                tmp = (uint32_t)val;
+                /* 
+                 * At first, cast to the integer of appropriate size,
+                 * then reject the most significant bits.
+                 */
+                tmp = (uint32_t)(uint64_t)val;
                 LGR_32_TO_NET(tmp, tmp_buf);
                 tmp_buf += sizeof(uint32_t);
 #else
